Add overflow detection strategy postcheck
This strategy first computes the value and then checks for overflow.  It
can only be applied to unsigned operations, because overflow at signed
arithmetics is undefined behaviour. In contrast, operations on unsigned
data types are guaranteed to wrap as if using the modulo operation.

The benchmark shows that the postcheck strategy is rather slow, so the
default streategy does not use it.