Add overflow_{likely,unlikely} functions
Add macros and functions that indicate the probability of an integer
overflow. By now, both versions, likely and unlikely, only use the
precheck strategy, but this will change in future versions.

Even when the programmer does not know the precise probability of
integer overflow conditions, these macros can still be useful for
de-optimization of error cases.

Type-generic macro versions of the likely/unlikely operations also
exist.

For testing and benchmarking, the pseudo-strategies likely and unlikely
are added so that the existing testcases and benchmarks need no
modifications.