Remove matrix benchmark
This benchmark is very slow, hardly comparable to the other benchmarks,
and some random environment effects influence the time results. For
instance, GCC decides to unroll some loops, but this depends on the data
type, and on the platform, etc.

As benchmarking takes a lot of time, and this particular benchmark does
not give any usable information, it is the best to remove it.

This also improves the benchmark comparison in the result plot, because
matrix is a lot slower than all other benchmarks.