Remove strategy largetype for unsigned subtraction
This strategy does not really make sense because only the <0 condition
is an overflow. Unfortunately, using a larger type does not prevent the
overflow, because all unsigned types wrap at 0.

This strategy had rather been accidentially correct, because the >MAX
condition happened to be equivalent to the <0 condition due to the
larger type. Thus, the largetype strategy is some kind of obfuscated
postcheck strategy for unsigned subtraction.

Therefore, it is probably the best idea to remove it. The default
strategy did not use it anyways.