Update internal function signatures
Update the function signatures for all internal functions.
 - make all parameters const
 - add restrict keyword to result pointer
The const keyword produces compiler errors on code modifications that
try to re-assign values to the parameters. This is a very bad idea
because it could break some assertions/assumptions. The restrict keyword
disallows aliasing of the result pointer with any other involved value.